Adds Supervisor to start new games
We want to keep track of running games.

As they're processes, the best way to do that is using a supervisor.
This way we can fine-tune their lifecycle.

To achieve this, we define a `GameSupervisor` that is able to spawn
new game processes on demand, adding them to the supervision tree.

Note the `:simple_one_for_one` strategy, where each child process is
handled independently.

We also add a convenience `start_new_game/0` function.