Adds a Game registry
- When a game is started, it's added to a registry
- When a game crashes/finishes, it's removed from the registry
- It's possible to get a list of all running games

We add here a custom, not particularly production-ready implementation
of a custom process registry.

The reason why we do that is to be able to find a process by uuid.

The `GameRegistry` module has to expose a specific set of functions,
hence why the beefy implementation. We just a add a `get_all/0` method
to retrieve a map with all games.