A player can join a game
We start introducing players, where a player is defined as the
combination of a uuid and a move.

Joining a game performs a nested state update, which proves to be
slightly inconvenient with Elixir 1.1 (compared to 1.0, where we were
able to leverage the `Access` protocol).

Note that we just care about the golden path: joining a game with less
than two players, with no acknowledgement to the caller. This is the
simplest possible case: we let the process crash once a 3rd player tries
to join. A solution to this is left as an exercise.