Players can make a move and finish the game
This is where Elixir's semantics really shine.

1. When a player makes a move, we don't perform any validation. If the
player is not already in the `players` map, using `Map.update!/3` will
throw an error, causing the game to crash. In other words, if a player
tries to make a move on a game they haven't joined, they will crash it,
but don't create an invalid state. This can be improved further, but
it's a good baseline.

2. Defining a valid move (i.e. only rock, paper or scissors) is just a
matter of defining a module attribute and a guard clause. Once again, we
let the process crash if any other move (like `fire` or `spock`) is
passed as an argument.

3. Finally, if a game is over we terminate the process with a `:normal`
reason. This will make sure that it's not restarted. Termination becomes
a great strategy to perform cleanup.