Adds GameDispatcher
So far we built the core game logic, it's now time to connect it to the
outside world.

We're going to use an event-based approach, i.e. every game will
broadcast specific events and its own state when necessary.

In the simplest possible implementation, we can use a GenEvent process,
knowing that this may become a performance bottleneck in the future.