Adds manager for supervised event handlers
OTP doesn't provide any built-in way to supervise a `GenEvent` handler,
so we have to write our own. The good news is that this is so generic
that it can be reused pretty much anywhere you like.

Our manager gets initialized with the dispatcher and a handler and
proceeds to start it and monitor it. In case the handler crashes, its
exit is trapped and another one is started.