Pipes game events to channel
In this commit we flesh out the channel API. While this could be
normally split in smaller chunks, it's presented as a single commit in
order to have a better global overview.

Some highlights:

1. We add some basic session handling to assign a uuid to a visitor.
This will be printed in the page template so that it can be used when
connecting to the channel.

2. The `UserSocket` receives the uuid and stores it in the socket state,
so that it's always available for all channel callbacks. This makes sure
that a player cannot make moves for other ones.

3. We implement all semantics for our game: start, join, move. Note that
we never return a direct response to the caller, as the evented
infrastructure we laid out beforehand will make sure that the state of
our games is always propagated correctly.

4. After joining, we send all existing games to the browser.